head: partially avoid an OOM abort with large byte counts
If the user-provided byte count is very large (and negative), avoid
allocating vectors with the byte count size.  Instead, allocate with
the default size and try to read the input data first.  Doing so
avoids an abort in the case where the file size is manageable even
though the given byte count is not.