posix/sh: rework builtin/function execution for redirection
While redirection is not yet supported, the framework is now in
place.  Due to creating custom UtilSetups for builtins/functions,
UtilSetup/UtilData needed to be reworked to use borrowed mutable
pointers rather than RefCell internally, which necessitated changes
in other utilities as well.  On the plus side, we no longer need
ensure the borrow checker's constraints are satisfied at runtime,
which should (hopefully) improve execution time slightly and
reduce the risk of an unintended panic.