posix/sh: implement cd (assuming we are removing util::actual_path)
I have made the assumption that we are just going to get rid of
util::actual_path() when working on this.  If we end up keeping it,
all that would really need to do to fix `cd` is avoid actually
using `env::set_current_dir()` and also make sure to use
`util::actual_path()` on anything that accesses the filesystem
(like checking if a path represents a directory).