posix/sh: store intermediate file descriptors
This *should* (I believe) prevent an issue where if the user
redirected a file descriptor A to file descriptor B (whose
underlying type owned some resource that would be closed on drop)
and then redirected file descriptor B to some other file
descriptor C, the original value of B would be dropped and
whatever resources it held would be closed.  Because the resource
was closed, operations that used A would then fail as well.